% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f3_cor_matrix.R
\name{cor_matrix_plot}
\alias{cor_matrix_plot}
\title{A nbaPlayers Function}
\usage{
cor_matrix_plot(data, year)
}
\arguments{
\item{data}{The input dataset}

\item{year}{The input value of year}
}
\description{
This function allows you to generate the correlation matrix for all numeric variables in the data for that year.
}
\examples{
cor_matrix_plot()
}
\keyword{cor_matrix}
\keyword{year,}
