% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f1_point_of_interest.R
\name{data_point_of_interest}
\alias{data_point_of_interest}
\title{A nbaPlayers Function}
\usage{
data_point_of_interest(data, year)
}
\arguments{
\item{data}{The input dataset}

\item{year}{The input value of year}
}
\description{
This function allows you to calculate a data point of interest in the dataset for some player in any given year.
}
\examples{
data_point_of_interest()
}
\keyword{year}
